refactor(types): align Package 0 scope with plan v2.6
- Remove sync.ts and mcp.ts (deferred to Package 4 and Integration)
- Move SearchFiltersSchema from mcp.ts to api.ts
- Add RerankResultSchema to api.ts (was missing)
- Update barrel exports to content, auth, api, storage only
