+++ .project/1/progress.md
@@ -1,31 +1,28 @@
 # Progress Log
 
 **Issue**: #1
-**Session**: 2026-01-31
+**Session**: 2026-02-03 (redo for spec v0.11)
 
 ## 5-Question Check
 
-1. What is the goal? Define shared Zod schemas and TypeScript types for all modules.
-2. What phase am I in? Implement
-3. What phases remain? PR, Review
-4. What was the last action? Ran type-check — passed zero errors
-5. What is the next action? Run /validate
+1. What is the goal? Redo Package 0 schemas to match plan v2.6 / spec v0.11 scope.
+2. What phase am I in? Complete.
+3. What phases remain? None — commit and merge.
+4. What was the last action? Ran type-check and /validate — both pass.
+5. What is the next action? Commit.
 
 ## Actions Taken
 
 | Time | Action | Result |
 |------|--------|--------|
-| 2026-01-31 | Created GitHub repo remote + pushed | success |
-| 2026-01-31 | Created 8 labels + 7 issues | success |
-| 2026-01-31 | Initialized .project/1/ with templates | success |
-| 2026-01-31 | Created content.ts, auth.ts | success |
-| 2026-01-31 | Created api.ts, storage.ts, sync.ts | success |
-| 2026-01-31 | Created mcp.ts, index.ts | success |
-| 2026-01-31 | Ran type-check | pass — zero errors |
-| 2026-01-31 | Validated with cf-validator | pass — all compliant |
-| 2026-01-31 | Committed setup to main, types to feature branch | success |
-| 2026-01-31 | Created PR #8 | success |
-| 2026-01-31 | Posted progress comment on issue #1 | success |
+| 2026-01-31 | Initial Package 0 implementation (old spec) | PR #8 merged |
+| 2026-02-03 | Deleted src/types/sync.ts (out of scope per plan v2.6) | success |
+| 2026-02-03 | Deleted src/types/mcp.ts (out of scope per plan v2.6) | success |
+| 2026-02-03 | Added SearchFiltersSchema to api.ts (moved from mcp.ts) | success |
+| 2026-02-03 | Added RerankResultSchema to api.ts (was missing) | success |
+| 2026-02-03 | Updated index.ts barrel exports (content, auth, api, storage only) | success |
+| 2026-02-03 | Ran type-check | pass — zero errors |
+| 2026-02-03 | Ran /validate | pass — types compliant, index.ts legacy issues out of scope |
 
 ## Error Log
 
@@ -37,4 +34,4 @@
 | Test | Status | Notes |
 |------|--------|-------|
 | npm run type-check | pass | zero errors |
-| Cloudflare validation | pass | no anti-patterns |
+| Cloudflare validation (types) | pass | 13 pass, 4 fail (all in index.ts legacy template, not types) |
